Removed Bits#readUtf8Char usage from JSON getter.
